Store original image files
When the :ORIGINAL-PATHNAME argument is specified to MAKE-STORE-IMAGE,
that argument's value is assumed to be the pathname of the original
image file.  Instead of writing out the CL-GD in-memory image, that
file is used to initialize the image blob.  That way, the image can
contain information that CL-GD does not handle (i.e. animations).